numpy & time Syntax Changes
numpy is no longer imported via pylab. It is now imported directly as np, and then np prefix has been added each time it is called. This step is intended to reduce namespace pollution.

In the Main Program Loop, pause(0.1) was changed to time.sleep(0.1) because "pause" is supported by the numpy or time libraries.